+++ src/graysvr/CContain.cpp
@@ -733,8 +733,8 @@ void CItemContainer::Trade_UpdateGold( DWORD platinum, DWORD gold )
 	bool bUpdateChar2 = pChar2->GetClient()->GetNetState()->isClientVersion(MINCLIVER_TOL);
 
 	// To prevent cheating, check if the char really have these gold/platinum values
-	INT64 iMaxValue = pChar1->m_virtualGold;
-	if ( gold + (platinum * 1000000000) > static_cast<UINT64>(iMaxValue) )
+	unsigned long long iMaxValue = pChar1->m_virtualGold;
+	if ( gold + (platinum * 1000000000) > iMaxValue )
 	{
 		gold = static_cast<DWORD>(iMaxValue % 1000000000);
 		platinum = static_cast<DWORD>(iMaxValue / 1000000000);
